REGOD-BE/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.10
asyncpg==0.29.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
python-multipart==0.0.6
alembic==1.12.1
docker==6.1.2
redis==5.0.1
PyJWT==2.8.0
email-validator==2.1.0
cryptography==41.0.7
websockets==12.0
pgmq==1.0.0